pyproject.toml
[project]
name = "repl-nix-cruisecompanion"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.11",
    "fastapi>=0.115.6",
    "flask-cors>=5.0.0",
    "flask>=3.1.0",
    "openai>=1.58.1",
    "psutil>=6.1.1",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
    "pymongo>=4.11.0",
    "python-dateutil>=2.8.2",
    "langchain-community",
    "langchain-anthropic",
    "langchain-experimental",
    "matplotlib",
    "langgraph",
    "langchain-openai",
    "sentence-transformers",
    "openpyxl",
    "pandas",
    "numpy",
    "scikit-image",
    "faiss-cpu",
    "langmem"]

[tool.poetry.dependencies]
python = "^3.8"
langgraph = "^1.0.0"  # Replace with the correct version if known
